occluder/requirements.txt
numpy>=1.17.3
#h5py>=2.9.0
# keras>=2.2.4
matplotlib>=3.1.1
scikit-learn>=0.21.3
pillow>=6.2.1

# Note that this is an older version (i.e. pre 2) of tensorflow
# tensorflow==1.14
